requirement.txt
# python = 3.12

streamlit
pandas
gspread
google-auth